NAS-121735: Move helptext and some code to external files
